tools/requirements.txt
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.9.0
babel==2.17.0
backoff==2.2.1
backrefs==5.8
beautifulsoup4==4.13.3
browser-use==0.1.40
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
defusedxml==0.7.1
distro==1.9.0
ghp-import==2.1.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jinja2==3.1.6
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain-anthropic==0.3.3
langchain-core==0.3.48
langchain-ollama==0.2.2
langchain-openai==0.3.1
langsmith==0.3.18
markdown==3.7
markdownify==0.14.1
markupsafe==3.0.2
mergedeep==1.3.4
mkdocs==1.6.1
mkdocs-get-deps==0.2.0
mkdocs-material==9.6.9
mkdocs-material-extensions==1.3.1
monotonic==1.6
ollama==0.4.7
openai==1.68.2
orjson==3.10.16
packaging==24.2
paginate==0.5.7
pathspec==0.12.1
platformdirs==4.3.7
playwright==1.51.0
posthog==3.21.0
pydantic==2.10.6
pydantic-core==2.27.2
pyee==12.1.1
pygments==2.19.1
pymdown-extensions==10.14.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pyyaml==6.0.2
pyyaml-env-tag==0.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
setuptools==78.1.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
tenacity==9.0.0
tiktoken==0.9.0
tqdm==4.67.1
typing-extensions==4.13.0
urllib3==2.3.0
watchdog==6.0.0
zstandard==0.23.0
